refactor(diagram): model 디렉토리용 type 파일 분리
